+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/TabsResponsiveToParent.test.jsx
@@ -19,7 +19,7 @@ class TabsResponsiveToParent extends React.Component {
 
   render() {
     return (
-      <div>
+      <div id="responsiveToParent">
         <button type="button" onClick={this.handleOnButtonClick}>Set Container Width to 600 px</button>
         <TabsTemplate responsiveTo="parent" containerClassName={this.state.containerClassName} />
       </div>