+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/Tabs/TabsWithIconsInMenu.test.jsx
@@ -10,26 +10,28 @@ import IconError from 'terra-icon/lib/icon/IconError';
 import TabContent from './TabContentTemplate';
 
 const TabsWithIconsInMenu = () => (
-  <Tabs id="tabsWithIconsInMenu" responsiveTo="parent">
-    <Tabs.Pane id="search" label="Search" icon={<IconSearch />} key="Search">
-      <TabContent label="Search Icon" />
-    </Tabs.Pane>
-    <Tabs.Pane id="required" label="Required" icon={<IconRequired style={{ color: 'red' }} />} isIconOnly key="Required">
-      <TabContent label="Required Icon, Icon Only" />
-    </Tabs.Pane>
-    <Tabs.Pane id="alert" label="Alert" icon={<IconAlert />} showIcon key="Alert">
-      <TabContent label="Alert Icon, Show in Tab and Menu" />
-    </Tabs.Pane>
-    <Tabs.Pane id="warning" label="Warning" icon={<IconWarning />} isIconOnly showIcon key="Warning">
-      <TabContent label="Warning Icon, Icon Only, Show in Tab and Menu" />
-    </Tabs.Pane>
-    <Tabs.Pane id="error" label="Error" icon={<IconError />} showIcon key="Error">
-      <TabContent label="Error Icon, Show in Tab and Menu" />
-    </Tabs.Pane>
-    <Tabs.Pane id="information" label="Information" isDisabled icon={<IconInformation />} showIcon key="Information">
-      <TabContent label="Information Icon, Disabled, Show in Tab and Menu" />
-    </Tabs.Pane>
-  </Tabs>
+  <div id="tabsWithIconsInMenu">
+    <Tabs responsiveTo="parent">
+      <Tabs.Pane id="search" label="Search" icon={<IconSearch />} key="Search">
+        <TabContent label="Search Icon" />
+      </Tabs.Pane>
+      <Tabs.Pane id="required" label="Required" icon={<IconRequired style={{ color: 'red' }} />} isIconOnly key="Required">
+        <TabContent label="Required Icon, Icon Only" />
+      </Tabs.Pane>
+      <Tabs.Pane id="alert" label="Alert" icon={<IconAlert />} showIcon key="Alert">
+        <TabContent label="Alert Icon, Show in Tab and Menu" />
+      </Tabs.Pane>
+      <Tabs.Pane id="warning" label="Warning" icon={<IconWarning />} isIconOnly showIcon key="Warning">
+        <TabContent label="Warning Icon, Icon Only, Show in Tab and Menu" />
+      </Tabs.Pane>
+      <Tabs.Pane id="error" label="Error" icon={<IconError />} showIcon key="Error">
+        <TabContent label="Error Icon, Show in Tab and Menu" />
+      </Tabs.Pane>
+      <Tabs.Pane id="information" label="Information" isDisabled icon={<IconInformation />} showIcon key="Information">
+        <TabContent label="Information Icon, Disabled, Show in Tab and Menu" />
+      </Tabs.Pane>
+    </Tabs>
+  </div>
 );
 
 export default TabsWithIconsInMenu;