+++ packages/terra-worklist-data-grid/tests/jest/Cell.test.jsx
@@ -1,0 +1,196 @@
+import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl } from 'terra-enzyme-intl';
+import Cell from '../../src/subcomponents/Cell';
+
+describe('Cell', () => {
+  it('verifies that only a row header cell (<th>) is created when isRowHeader prop is true', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        isTabStop={false}
+        isSelected={false}
+        isRowHeader
+        onCellSelect={jest.fn}
+        height="50px"
+      >
+        Data in cell
+      </Cell>,
+    ).dive();
+
+    expect(wrapper.find('th')).toHaveLength(1); // Th when row header.
+    expect(wrapper.find('td')).toHaveLength(0); // There should be no th since this is row header.
+    expect(wrapper).toMatchSnapshot();
+
+    const cellContent = wrapper.find('th').find('div');
+    expect(cellContent).toHaveLength(1);
+    expect(cellContent.text()).toEqual('Data in cell');
+  });
+
+  it('verifies that only a regular cell(<td>) is created when isRowHeader prop is false', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        isTabStop={false}
+        isSelected={false}
+        isRowHeader={false}
+        onCellSelect={jest.fn}
+      >
+        Data in cell
+      </Cell>,
+    ).dive();
+
+    const normalCell = wrapper.find('td');
+    expect(normalCell).toHaveLength(1); // Th when non row header.
+    expect(wrapper.find('th')).toHaveLength(0); // There should be no td since this is a row header.
+
+    const cellContent = normalCell.find('div');
+    expect(cellContent).toHaveLength(1);
+    expect(cellContent.text()).toEqual('Data in cell');
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies that the cell rendered is marked selectable when isSelectable prop is true', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        isSelectable
+        onCellSelect={jest.fn}
+      >
+        Data in cell
+      </Cell>,
+    ).dive();
+
+    const selectableCell = wrapper.find('td.selectable');
+    expect(selectableCell).toHaveLength(1); // Cell should have been styled selectable.
+
+    const cellContent = selectableCell.find('div');
+    expect(cellContent).toHaveLength(1);
+    expect(cellContent.text()).toEqual('Data in cell');
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies that the cell rendered is marked blank when the cell content is empty', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        onCellSelect={jest.fn}
+      />,
+    ).dive();
+
+    const blankCell = wrapper.find('td.blank');
+    expect(blankCell).toHaveLength(1);
+
+    const cellContent = blankCell.find('div');
+    expect(cellContent).toHaveLength(0);
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies that a selected cell has the correct styles when isSelected prop is true', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        ariaLabel="Some Label Here"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        isSelected
+        onCellSelect={jest.fn}
+      >
+        Data in cell
+      </Cell>,
+    ).dive();
+
+    const selectedCell = wrapper.find('td.selected');
+    expect(selectedCell).toHaveLength(1); // Cell should have been styled selected.
+
+    expect(wrapper.find('td').prop('aria-selected')).toBe(true);
+    expect(wrapper.find('td').prop('aria-label')).toBe('Some Label Here');
+
+    const cellContent = selectedCell.find('div');
+    expect(cellContent).toHaveLength(1);
+    expect(cellContent.text()).toEqual('Data in cell');
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies that a cell has the correct styles and no content when isMasked prop is true', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        ariaLabel="Some Label Here"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        isMasked
+        onCellSelect={jest.fn}
+      >
+        Data in cell
+      </Cell>,
+    ).dive();
+
+    const maskedCell = wrapper.find('td.masked');
+    expect(maskedCell).toHaveLength(1);
+
+    expect(wrapper.find('td').prop('aria-label')).toBe('Terra.worklistDataGrid.maskedCell');
+
+    const cellContent = maskedCell.find('div');
+    expect(cellContent).toHaveLength(0); // Masked cell should have no content.
+
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('verifies mask takes precedence when cell is masked, selectable and selected', () => {
+    const wrapper = shallowWithIntl(
+      <Cell
+        rowId="RowID"
+        columnId="ColumnId"
+        ariaLabel="Some Label Here"
+        rowIndex={1}
+        columnIndex={2}
+        key="key"
+        isSelected
+        isMasked
+        isSelectable
+        onCellSelect={jest.fn}
+      >
+        Data in cell
+      </Cell>,
+    ).dive();
+
+    const maskedCell = wrapper.find('td.masked');
+    expect(maskedCell).toHaveLength(1);
+
+    expect(maskedCell.prop('aria-label')).toBe('Terra.worklistDataGrid.maskedCell');
+
+    const cellContent = maskedCell.find('div');
+    expect(cellContent).toHaveLength(0); // Masked cell should have no content.
+
+    // Verify other styles not applied.
+    expect(wrapper.find('td.selected')).toHaveLength(0); // Cell should not be styled selected since it is masked.
+    expect(wrapper.find('td.selectable')).toHaveLength(0); // Cell should not be styled selectable since it is masked.
+
+    expect(wrapper).toMatchSnapshot();
+  });
+});
+