+++ packages/terra-worklist-data-grid/tests/jest/ColumnHeaderCell.test.jsx
@@ -5,10 +5,10 @@ import IconUp from 'terra-icon/lib/icon/IconUp';
 import IconDown from 'terra-icon/lib/icon/IconDown';
 import IconError from 'terra-icon/lib/icon/IconError';
 import { shallowWithIntl } from 'terra-enzyme-intl';
-import ColumnHeaderCell from '../../src/ColumnHeaderCell';
+import ColumnHeaderCell from '../../src/subcomponents/ColumnHeaderCell';
 
 describe('ColumnHeaderCell', () => {
-  it('render default column header cell', () => {
+  it('renders a default column header cell', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -16,9 +16,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
       />,
     ).dive();
 
@@ -27,18 +28,18 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
     expect(headerContainer.children()).toHaveLength(1);
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
 
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('render column header cell with ascending sort', () => {
+  it('renders a column header cell with ascending sort', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -47,9 +48,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
       />,
     ).dive();
 
@@ -58,19 +60,19 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
     expect(headerContainer.find(IconUp)).toHaveLength(1);
 
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('render column header cell with descending sort', () => {
+  it('renders a column header cell with descending sort', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -79,9 +81,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
       />,
     ).dive();
 
@@ -90,19 +93,19 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('descending');
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
     expect(headerContainer.find(IconDown)).toHaveLength(1);
 
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('render column header cell with error', () => {
+  it('renders a column header cell with error', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -111,9 +114,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
       />,
     ).dive();
 
@@ -122,19 +126,19 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBeUndefined();
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
     expect(headerContainer.find(IconError)).toHaveLength(1);
 
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('render column header cell with ascending sort and error', () => {
+  it('renders a column header cell with ascending sort and error', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -144,9 +148,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
       />,
     ).dive();
 
@@ -155,20 +160,20 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
     expect(headerContainer.find(IconUp)).toHaveLength(1);
     expect(headerContainer.find(IconError)).toHaveLength(1);
 
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('render column header cell with onColumnSelect callback', () => {
+  it('renders a column header cell with onColumnSelect callback', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -180,9 +185,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
         onColumnSelect={mockClick}
       />,
     ).dive();
@@ -192,21 +198,21 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
-    expect(columnHeader.props().onClick).toBeDefined();
+    expect(columnHeader.props().onMouseDown).toBeDefined();
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
     expect(headerContainer.find(IconUp)).toHaveLength(1);
     expect(headerContainer.find(IconError)).toHaveLength(1);
 
     expect(wrapper).toMatchSnapshot();
   });
 
-  it('render column header cell with onColumnSelect callback but not selectable', () => {
+  it('renders a column header cell with onColumnSelect callback but not selectable', () => {
     const column = {
       id: 'Column-0',
       displayName: ' Vitals',
@@ -219,9 +225,10 @@ describe('ColumnHeaderCell', () => {
 
     const wrapper = shallowWithIntl(
       <ColumnHeaderCell
-        column={column}
-        width="100px"
+        columnIndex={0}
+        width={100}
         headerHeight="150px"
+        {...column}
         onColumnSelect={onColumnSelect}
       />,
     ).dive();
@@ -231,14 +238,14 @@ describe('ColumnHeaderCell', () => {
     expect(columnHeader.key()).toBe('Column-0');
     expect(columnHeader.props().role).toBe('columnheader');
     expect(columnHeader.props().scope).toBe('col');
-    expect(columnHeader.props().tabIndex).toBe('-1');
+    expect(columnHeader.props().tabIndex).toEqual(-1);
     expect(columnHeader.props()['aria-sort']).toBe('ascending');
-    expect(columnHeader.props().onClick).toBeUndefined();
+    expect(columnHeader.props().onMouseDown).toBeUndefined();
     expect(columnHeader.props().style.width).toBe('100px');
     expect(columnHeader.props().style.height).toBe('150px');
 
-    const headerContainer = columnHeader.find('.header-container');
-    expect(headerContainer.find('span[role="button"]').text().trim()).toBe('Vitals');
+    const headerContainer = columnHeader.find('.header-container[role="button"]');
+    expect(headerContainer.find('span').text().trim()).toBe('Vitals');
     expect(headerContainer.find(IconUp)).toHaveLength(1);
     expect(headerContainer.find(IconError)).toHaveLength(1);