+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Locked `uuid` dependency to `7.0.3` for consistency across Terra packages.
+
 ## 1.71.0 - (June 12, 2023)
 
 * Changed